Replace parser with devices
