--- parser/requirements-dev.in
@@ -1,9 +1,0 @@
-autoflake
-autopep8
-black
-flake8
-isort
-pep8-naming
-pytest
-pytest-cov
-pytest-flask
