--- parser/requirements-dev.txt
@@ -1,40 +1,0 @@
-#
-# This file is autogenerated by pip-compile
-# To update, run:
-#
-#    pip-compile requirements-dev.in
-#
-appdirs==1.4.4            # via black
-attrs==20.2.0             # via pytest
-autoflake==1.4            # via -r requirements-dev.in
-autopep8==1.5.4           # via -r requirements-dev.in
-black==20.8b1             # via -r requirements-dev.in
-click==7.1.2              # via black, flask
-coverage==5.3             # via pytest-cov
-flake8-polyfill==1.0.2    # via pep8-naming
-flake8==3.8.4             # via -r requirements-dev.in, flake8-polyfill
-flask==1.1.2              # via pytest-flask
-iniconfig==1.1.1          # via pytest
-isort==5.6.4              # via -r requirements-dev.in
-itsdangerous==1.1.0       # via flask
-jinja2==2.11.2            # via flask
-markupsafe==1.1.1         # via jinja2
-mccabe==0.6.1             # via flake8
-mypy-extensions==0.4.3    # via black
-packaging==20.4           # via pytest
-pathspec==0.8.0           # via black
-pep8-naming==0.11.1       # via -r requirements-dev.in
-pluggy==0.13.1            # via pytest
-py==1.9.0                 # via pytest
-pycodestyle==2.6.0        # via autopep8, flake8
-pyflakes==2.2.0           # via autoflake, flake8
-pyparsing==2.4.7          # via packaging
-pytest-cov==2.10.1        # via -r requirements-dev.in
-pytest-flask==1.0.0       # via -r requirements-dev.in
-pytest==6.1.2             # via -r requirements-dev.in, pytest-cov, pytest-flask
-regex==2020.10.28         # via black
-six==1.15.0               # via packaging
-toml==0.10.1              # via autopep8, black, pytest
-typed-ast==1.4.1          # via black
-typing-extensions==3.7.4.3  # via black
-werkzeug==1.0.1           # via flask, pytest-flask
