--- parser/requirements-prod.in
@@ -1,1 +1,0 @@
-uwsgi<3
