--- parser/requirements-prod.txt
@@ -1,7 +1,0 @@
-#
-# This file is autogenerated by pip-compile
-# To update, run:
-#
-#    pip-compile requirements-prod.in
-#
-uwsgi==2.0.19.1           # via -r requirements-prod.in
