--- parser/requirements.txt
@@ -1,19 +1,0 @@
-#
-# This file is autogenerated by pip-compile
-# To update, run:
-#
-#    pip-compile
-#
-certifi==2020.6.20        # via -r requirements.in, fvhiot
-click==7.1.2              # via flask
-flask==1.1.2              # via fvhiot
-git+https://github.com/ForumViriumHelsinki/FVHIoT-python.git#egg=fvhiot  # via -r requirements.in
-itsdangerous==1.1.0       # via flask
-jinja2==2.11.2            # via flask
-kafka-python==2.0.2       # via fvhiot
-markupsafe==1.1.1         # via jinja2
-msgpack==1.0.0            # via fvhiot
-python-dateutil==2.8.1    # via -r requirements.in
-pytz==2020.4              # via -r requirements.in
-six==1.15.0               # via python-dateutil
-werkzeug==1.0.1           # via flask
