--- parser/setup.cfg
@@ -1,26 +1,0 @@
-[pep8]
-max-line-length = 120
-ignore = E309
-
-[flake8]
-max-line-length = 120
-max-complexity = 10
-
-[tool:pytest]
-norecursedirs = bower_components node_modules .git venv*
-doctest_optionflags = NORMALIZE_WHITESPACE IGNORE_EXCEPTION_DETAIL ALLOW_UNICODE
-
-[coverage:run]
-branch = True
-omit = *migrations*,*site-packages*,*venv*,*tests*
-
-[isort]
-known_first_party=app
-multi_line_output=3
-include_trailing_comma=True
-force_grid_wrap=0
-use_parentheses=True
-line_length=88
-
-[pydocstyle]
-ignore=D100,D104,D105,D200,D203,D400
